+++ RELEASENOTES.md
@@ -11,6 +11,10 @@
     *   Fix a bug where an assertion would fail if the player started to buffer
         an ad media period before the ad URI was known then an ad state update
         arrived that didn't set the ad URI.
+*   DRM:
+    *   Re-use the previous `DrmSessionManager` instance when playing a playlist
+        (if possible)
+        ([#8523](https://github.com/google/ExoPlayer/issues/8523)).
 
 ### 2.13.0 (2021-02-04)